+++ src/app/cli.py
@@ -6,12 +6,10 @@ from src.__version__ import __version__
 def parse_args():
     """Parse command-line arguments and run the application.
 
-    Sets up argument parser with options for verbose mode and Homebrew updates,
+    Sets up argument parser with options for Homebrew updates,
     parses the command-line arguments, and invokes the main update process.
     """
 
-    new_kernel = True
-    verbose = False
     brew = False
 
     parser = argparse.ArgumentParser(
@@ -24,11 +22,6 @@ def parse_args():
         version=f"%(prog)s {__version__}"
     )
     parser.add_argument(
-        "--verbose", "-l", "--log",
-        action="store_true",
-        help="Show detailed output during update process"
-    )
-    parser.add_argument(
         "--brew", "-b",
         action="store_true",
         help="Update Homebrew packages (if installed)"
@@ -36,14 +29,12 @@ def parse_args():
 
     args = parser.parse_args()
 
-    # Extract arguments into boolean variables
-    verbose = args.verbose
     brew = args.brew
 
     print("\n--- Tuxgrade - Linux System Updater ---\n")
 
     # Run the main update process
-    app.run(verbose, brew)
+    app.run(brew)
 
     print("\n--- System Upgrade finished ---\n")