+++ src/distros/arch_distro.py
@@ -13,24 +13,23 @@ class ArchDistro(GenericDistro):
     (Snap, Flatpak, Homebrew).
     """
 
-    def update(self, verbose, brew):
+    def update(self, brew):
         """Perform system updates for Arch distributions.
 
         Currently delegates to the parent GenericDistro class to update
         common package managers (Snap, Flatpak, Homebrew).
 
         Args:
-            verbose: If True, show detailed output; if False, show minimal output with spinners.
             brew: If True, include Homebrew package updates.
         """
 
-        cli_print_utility.print_header("Update Arch Packages", verbose)
+        cli_print_utility.print_header("Update Arch Packages")
 
         if(paru.avail()):
-            cli_print_utility.print_output(paru.update_paru, verbose, "Updating packages with Paru")
+            cli_print_utility.print_output(paru.update_paru)
         elif(yay.avail()):
-            cli_print_utility.print_output(yay.update_yay, verbose, "Updating packages with Yay")
+            cli_print_utility.print_output(yay.update_yay)
         else:
-            cli_print_utility.print_output(pacman.update_pacman, verbose, "Updating packages with Pacman")
+            cli_print_utility.print_output(pacman.update_pacman)
 
-        super().update(verbose, brew)
+        super().update(brew)